micro-services/event-logging-controller/requirements.txt
confluent-kafka==1.8.2
kafka-python==2.0.1
envyaml
Flask==2.0.1
requests==2.25.1
SQLAlchemy
sqlalchemy-utils
marshmallow
prometheus-client==0.11.0
/modules/camunda
mysqlclient
schedule